+++ native/src/main/java/io/ballerina/stdlib/graphql/runtime/schema/TypeFinder.java
@@ -24,6 +24,7 @@ import io.ballerina.runtime.api.types.ArrayType;
 import io.ballerina.runtime.api.types.Field;
 import io.ballerina.runtime.api.types.MapType;
 import io.ballerina.runtime.api.types.MethodType;
+import io.ballerina.runtime.api.types.Parameter;
 import io.ballerina.runtime.api.types.RecordType;
 import io.ballerina.runtime.api.types.RemoteMethodType;
 import io.ballerina.runtime.api.types.ResourceMethodType;
@@ -119,12 +120,11 @@ public class TypeFinder {
     }
 
     private void getInputTypesFromMethod(MethodType methodType) {
-        Type[] inputTypes = methodType.getType().getParameterTypes();
-        for (Type type : inputTypes) {
-            if (type.getTag() == TypeTags.RECORD_TYPE_TAG) {
-                getInputObjectSchemaType(type);
+        for (Parameter parameter : methodType.getParameters()) {
+            if (parameter.type.getTag() == TypeTags.RECORD_TYPE_TAG) {
+                getInputObjectSchemaType(parameter.type);
             } else {
-                getSchemaTypeFromBalType(type);
+                getSchemaTypeFromBalType(parameter.type);
             }
         }
     }